requirements.txt
openai
python-dotenv
fastapi
uvicorn
pypdf
faiss-cpu
numpy
